fix: replace inner PapayaMap with locked HashMap
The current build regularly segfaults presumable due to tracking
changes. The usual suspect, rkyv, is most likely not the culprit
this time so papaya is next in line. Unsure if replacing these
inner maps with RwLock'd HashMaps fixes the issue, though. If not,
the next attempt could be to replace the outer PapayaMap too. If
that's also not the cause, we'll have to search some more.